+++ libraries/MAX31855/examples/max31855_demo3/max31855_demo3.ino
@@ -1,7 +1,7 @@
 //
 //    FILE: max31855_demo3.ino
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.03
+// VERSION: 0.1.4
 // PURPOSE: thermocouple lib demo application
 //    DATE: 2014-01-02
 //     URL:
@@ -33,7 +33,7 @@ void setup()
   Serial.println(stop - start);
 
   start = micros();
-  double t1 = tc.getTemperature();
+  float t1 = tc.getTemperature();
   stop = micros();
   Serial.print("getTemperature:\t");
   Serial.println(stop - start);
@@ -55,8 +55,8 @@ void loop()
   // this loop returns multiples of about 73mSec (counter multiples of ~143)
   // so the # measurements per second is about 14?
   uint32_t counter = 0;
-  double t1 = tc.getTemperature();
-  double t2 = t1;
+  float t1 = tc.getTemperature();
+  float t2 = t1;
 
   uint32_t start = micros();
   while (t2 == t1)
